build: Update Gradle, AGP, and dependencies for AGP 9.0 compatibility
This commit updates the project to Gradle 9.1 and AGP 9.0.1, including necessary configuration adjustments for compatibility with existing plugins like SQLDelight.

Key changes include:
- **Build Infrastructure**:
    - Updated **Gradle** from 8.14.3 to 9.1.0.
    - Updated **Android Gradle Plugin (AGP)** from 8.13.2 to 9.0.1.
    - Added compatibility flags in `gradle.properties` (`android.disallowKotlinSourceSets=false`, `android.newDsl=false`, and `android.r8.proguardAndroidTxt.disallowed=false`) to support SQLDelight 2.2.x on AGP 9.
    - Removed redundant `kotlin-android` plugin applications across `app`, `data`, and root `build.gradle.kts` files.
- **Dependency Updates**:
    - Updated `kspVersion` from "2.3.4" to "2.3.6" in `libs.versions.toml`.
- **Code Refactoring**:
    - **UI**: Migrated deprecated `MenuAnchorType` to `ExposedDropdownMenuAnchorType` in `Select.kt`.
    - **Dashboard**: Simplified `totalPending` calculation in `DashboardScreen.kt` by removing an unnecessary `toLong()` conversion.
- **Module Configuration**:
    - Enabled `resValues` in the `data` module's `buildFeatures`.

+++ app/src/main/kotlin/com/emm/hello/newfeatures/dashboard/DashboardScreen.kt
@@ -109,7 +109,7 @@ fun DashboardScreen(
         ) {
             // ── Resumen de sesión ────────────────────────────────────────────
             item {
-                val totalPending = state.decks.sumOf { it.cardsCount.toLong() }
+                val totalPending = state.decks.sumOf { it.cardsCount }
                 AnimatedVisibility(
                     visible = !state.isLoading && totalPending > 0,
                     enter = fadeIn(),
